Add Atom packages and themes data
This commit adds my list of current packages and themes I use for
Atom. It also adds a helper script that both updates these packages
and themes files, and installs them.

The install bit could be done with the base `apm` command, but, after
some testing, it looks like it will install all packages regardless of
whether they are already installed, and there's no flag or command
option to avoid this behaviour, so I've had to wrap that behaviour
into the script for the moment.