Add a basic profile to system-config.yaml
The sater-manjaro profile might not be completely suitable for the
manjaro-setup/bootstrap.sh script, since it has a few files and
changes that rely completely on the sater-manjaro hardware and setup.

However, some other system files, like the fonts/local.conf file, that
sets Fantasque Sans Mono as the default monospace font for the system,
are still useful and needed.

Therefore, the basic profile will have those files that are mostly
generic, and that can and should be installed on whatever other system
I might use.