Add 71 and 79 default column guides to Atom
On some specific languages, like Python, Rust, or JavaScript I have
custom column guides. But by default, I usually keep 71 and 79. This
is to match Git's 72 commit body message length, and 80 which is often
the default on some languages and console stuff.

I use N-1 columns so that in a full line, the last character is always
a newline. This is mostly just I thing I like since I also keep
non-printable characters visible so I can easily see, on a full line,
that the only possible character after the guide is a newline.