Add hacky way to remove Atom's packages' versions
By default the `apm list` command lists the packages' versions too.
This is inconvenient for me since I want to always install the latest
version on any system I might use. It's also inconvenient because
whenever I try to update the `atom-*.list` files, there might be
version changes which dirty the commit diffs.

Just to note, these versions are stripped when installing packages. I
use `cut` to remove the version and just run `apm install
package-name`. However, that still leaves me with the issue of dirty
commit diffs.

I recently sent a PR to `apm` to add a flag that would allow listing,
or not, the packages' versions: https://github.com/atom/apm/pull/869
But it's received apparently no attention, so until it does, if it
does, I've decided to write some Bash magic to strip the version
numbers before storing the packages in the `atom-*.list` files. When
that PR is accepted, if it is, I'll be able to rid myself of the
changes in this commit.