Add kitty-grab, temporarily my fork version
kitty-grab is a kitten for Kitty, the terminal I currently use. It
allows to go into a mode similar to Vim's shift+v (Visual mode) that
allows to select text. Only that instead of doing it inside of Vim, it
does it on the whole terminal window, allowing me to select and copy
text with my keyboard rather than with my mouse.

As for why I'm temporarily using my own fork of it, rather than the
original package: yurikhan/kitty-grab#8