Use community Atom rather than AUR
The reason why I started using the AUR version is that when I was
setting up my first Manjaro system the community version was out of
date, and had been that way for weeks, with no sign of the maintainer
anywhere, so I went to the AUR version which was up to date. But now
it seems the community version is up to date, so I'd rather use that
(and it should take less time than building the AUR version too).