Pull changes in dotdrop to allow optional configs
After opening an issue on the dotdrop repository, and talking it with
the creator, he added the code to make it possible to have optional
variables/actions/config external files. However, it's not still
merged to the master branch so for now I'm still using my fork so I
can keep on working on my things.

Once that branch of dotdrop is merged back into master I'll get rid of
my fork and go back to using the original dotdrop.