Install my fork of i3exit on Manjaro boostrap
This fork uses xsecurelock instead of blurlock/i3lock. Blurlock is a
wrapper of i3lock by the Manjaro team, but it's far too slow. Mainly
because it takes and manipulates PNG images for the background. JPGs
would take less time overall, making the thing lock faster, but still.

Looking around I found xsecurelock and I like it so I'm switching to
that.