Set Docker to use pass as credentials manager
Apparently if you want to upload your image to DockerHub you need to
first login, and then, as a logged in user, in completely different
steps, upload and do any other thing that requires permissions.

That's already an annoyance since most registry tools I've used do not
have separate steps for login. In fact, they don't have steps for
login at all. When you try to do something that requires permissions,
it will ask them for it and then forget them. Think Pypi (setup.py
upload or twine), they both ask for your login only when uploading,
and then promptly forget it.

Instead, Docker not only requires a separate step, but also seems to
save that login information in clear text in `~/.docker/config.json`.
There's no way to work with SSH as there is with Gitlab/Github, so the
only options seem these `docker-credentials-*` helper packages they
develop and provide.

Pass seems pretty nice and standard, although I've yet to use it. So
far I've only used it for this Docker login and it seems to work just
fine, and Docker won't save my stuff in plain text for everyone to
see.