Fix i3 shortcut to ranger
Seems like ranger might need to be called from a terminal to start at
all. I'm suspecting dmenu calls everything from a terminal as well,
which is why it was working with dmenu (even though with the wrong
terminal) but not with the i3 shortcut.

It seems simply calling ranger through your terminal of choice in the
shortcut will do the trick. I'm using i3-sensible-terminal here to
call it, but because of previous commits where I set the TERMINAL
environment variable to kitty it's simply a thin wrapper for kitty in
this system.