i3: Assign some workspaces to specific monitors
This depends on any given profile, so there's no default setting for
the basic-manjaro profile. But for instance, on my home setup I mostly
have workspaces 1 and 9 on my main monitor, and workspace 2 on the
left monitor.

That's my startup setup. From then, throughout my use in the session I
might spawn workspaces 3 and 4 on any of the two monitors depending on
my needs.

Same thing goes for my work setup. I have some workspaces I always
want assigned to specific monitors because of the way I startup my
system. On there, it's even more useful when coupled with autorandr,
because after every undock all workspaces would switch to the main
laptop monitor, and then stay that way when docked again. Autorandr,
when coupled with this setting, allows for workspaces to be quickly
switched back to my usual monitors when docked again.