Swith to my fork of kitty-themes to fix Afterglow
You can see the changes in my fork of the kitty-themes repo, but as of
writing this, my issue was with the Afterglow theme having
non-sensible colours for selections. The selection would have no
contrast, meaning the foreground was dark on a gray background, making
reading what was selected pretty much impossible.