Configure VSCode
This commit adds more settings configurations, some keybindings, and a
few styles.

It also adds a f_vscode_packages dotfile. It's managed manually, and
it's not actually used by VSCode in any way. It's simply a dotfile for
me to keep an up to date list of my current packages. By keping it as
a dotfile, I can use profile variables to keep different packages in
different environments, and I can automate installing all the packages
in the file by using a post action.