Code: Add Rewrap extension
This extension allows automatically wrapping the (default) first
ruler. This is especially useful for me for writing Markdown documents
and anything that might not have its own specific formatter. When
that's the case, I tend to just stick to 80 chars per line. But it
gets annoying to manually have to newline myself, when it could simply
be handled by a machine.

The diff in the language rulers comes as direct result of this
package. You can change what ruler it will use to wrap at runtime with
Alt+Q, however, it uses the first ruler by default. But in my case,
I'd order by rulers in an ascending manner, with anything but the last
being guidelines and/or reminders, so having to press Alt+Q on each
file all the time to get it to use the last one all the time was
annoying. So I just moved the last one (so furthest) at the beginning
so it will be picked by this extension and used to wrap.