Add more options to yay to hopefully stop rebuilds
I'm having quite a few issues with Yay not noticing it already has the
packages I want installed, and so it downloads them again, it builds
them again, and it tries to install them once again. Honestly quite a
pain to deal with. Maybe these new arguments will help. I doubt it.