Try to validate and use hostname as dotdrop profile
This makes use of yq (which in turn makes use of jq) to parse the YAML
config files of the normal and sudo/system dotfiles.

Overall the logic is as follows: A default profile and config path for
each dotfile type is set. The default profile name for both right now
is basic-manjaro. Now, at runtime, the hostname is fetched, and if
it's a valid profile name in the dotdrop configuration YAML, then the
default will be overwritten with the hostname.

This was done because I often run this script to update my work system
with new dotfiles and new changes. And I'd often forget to specify the
custom work profile.

Note: Specifying custom profile names and/or paths through the CLI
options will overwrite the default and hostname fetched by this new
change.