Fix bootstrap code failing because of yq dep
The Manjaro bootstrap script makes use of yq to parse the YAML Dotdrop
config files, and see if the current hostname is a defined key in any
of the two. In case it is, it overwrites (unless explicitly passed
through the CLI) the default profile name used with the hostname.

But this code would fail in some cases. Mostly in cases where it's run
on a new system that doesn't have yq installed, because the function
doing all that hostname checking makes use of yq, but yq wouldn't be
installed until much later, meaning this function would fail.