Configure bootstrap.sh to use my own repo
This is because some packages that were previously AUR dependencies,
and so build with yay, can now be fetched from my own repository. As
well as some dependencies I built and installed manually through
GitHub, which are also now hosted on my repository.

Now I just run `pacman -Syy` to update the repos once the
`pacman.conf` file is set through the system dotfiles, since it
contains the definition for my repository.