Update rofi theme to rofi v1.6.0
This new release has a few breaking changes. I haven't been able to
actually figure out the exact changes that made the listview
(applications) break and stop displaying elements, but I wasn't able
to figure it out.

Getting rid of most custom boxes seems to have fixed the issue.

Added a light background to the mode-switcher since it seems the
text-color for the mode-switcher element (previously sidebar) can't be
changed, which means it has the default black color, which doesn't go
well with the default background of this theme.

Also renamed this theme to `dmenu-lb`.