Enable gpg-agent with longer key caches
This is useful later on for keychain, as it will just use the gpg-agent
configuration it has in the system. home-manager seems to only write the
configuration if you enable the agent, but that's not a big deal for me
since keychain will search for running agents before spawning its own.